+++ components/composer/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - [Composer] focus the HTML editor textbox when a formatting option is clicked [#342](https://github.com/nylas/components/pull/342)
 - [Composer] Header should reactively update with the change in the subject [#367](https://github.com/nylas/components/pull/367)
+- [Composer] Fix bug where unable to tab into contenteditable area in Firefox [#400](https://github.com/nylas/components/pull/400)
 
 # v1.1.6 (Unreleased)